+++ zebra-chain/src/amount/tests.rs
@@ -1,3 +1,4 @@
 //! Tests for amounts
 
 mod prop;
+mod vectors;
