+++ zebra-state/src/service/check/utxo.rs
@@ -3,7 +3,7 @@
 use std::collections::{HashMap, HashSet};
 
 use zebra_chain::{
-    block,
+    amount, block,
     transparent::{self, CoinbaseSpendRestriction::*},
 };
 
@@ -225,8 +225,8 @@ pub fn remaining_transaction_value(
     prepared: &PreparedBlock,
     utxos: &HashMap<transparent::OutPoint, transparent::Utxo>,
 ) -> Result<(), ValidateContextError> {
-    for transaction in prepared.block.transactions.iter() {
-        // This rule does not apply to coinbase transactions.
+    for (tx_index_in_block, transaction) in prepared.block.transactions.iter().enumerate() {
+        // TODO: check coinbase transaction remaining value (#338, #1162)
         if transaction.is_coinbase() {
             continue;
         }
@@ -236,15 +236,33 @@ pub fn remaining_transaction_value(
         match value_balance {
             Ok(vb) => match vb.remaining_transaction_value() {
                 Ok(_) => Ok(()),
-                Err(_) => Err(ValidateContextError::InvalidRemainingTransparentValue {
-                    transaction_hash: transaction.hash(),
-                    in_finalized_state: false,
-                }),
+                Err(amount_error @ amount::Error::Constraint { .. })
+                    if amount_error.invalid_value() < 0 =>
+                {
+                    Err(ValidateContextError::NegativeRemainingTransactionValue {
+                        amount_error,
+                        height: prepared.height,
+                        tx_index_in_block,
+                        transaction_hash: prepared.transaction_hashes[tx_index_in_block],
+                    })
+                }
+                Err(amount_error) => {
+                    Err(ValidateContextError::CalculateRemainingTransactionValue {
+                        amount_error,
+                        height: prepared.height,
+                        tx_index_in_block,
+                        transaction_hash: prepared.transaction_hashes[tx_index_in_block],
+                    })
+                }
             },
-            Err(_) => Err(ValidateContextError::InvalidRemainingTransparentValue {
-                transaction_hash: transaction.hash(),
-                in_finalized_state: false,
-            }),
+            Err(value_balance_error) => {
+                Err(ValidateContextError::CalculateTransactionValueBalances {
+                    value_balance_error,
+                    height: prepared.height,
+                    tx_index_in_block,
+                    transaction_hash: prepared.transaction_hashes[tx_index_in_block],
+                })
+            }
         }?
     }
